Sprint3 (#83)
* Izlistavanje rata za kredite
I zapocet Cron Job

* Dodato metode za vracanje broja preostalih rata

* Dodato metode za vracanje broja preostalih rata

* Cron job finished test in progress

* LoanServiceTest modification

* LoanPayment test in chunks

* Popravljeno nesto oko testova

* Prepravljeno

---------

+++ banking-service/src/main/java/com/banka1/banking/dto/request/CreateLoanDTO.java
@@ -15,7 +15,7 @@ public class CreateLoanDTO {
     private Double nominalRate;
     private Double effectiveRate;
     private Double loanAmount;
-    private Double duration;
+    private Integer duration;
     private Long allowedDate;
     private Double monthlyPayment;
     private CurrencyType currencyType;
